add restriction in delete

+++ controllers/option-controller.js
@@ -1,4 +1,5 @@
 const Option = require('../models/option');
+const Question = require('../models/question');
 
 module.exports.delete = async function(req, res){
     try {
@@ -7,6 +8,11 @@ module.exports.delete = async function(req, res){
         if(!option){
             return res.status(400).json({
                 error: "Option not found"
+            })
+        }
+        if(option.votes>0){
+            return res.status(400).json({
+                error: "Can't delete this option as it has votes"
             })
         }
         await option.deleteOne();
@@ -30,6 +36,8 @@ module.exports.vote = async function(req, res){
             })
         }
         await Option.updateOne({_id: option._id}, {$set:{votes :option.votes+1}});
+        console.log(option);
+        await Question.updateOne({_id: option.question}, {$set:{vote :true}});
         return res.status(200).json({
             message: "Your vote added successfully"
         });
